+++ Assets/Tests/Failbetter/Reader.cs
@@ -1,21 +1,38 @@
+using Ink;
+using Ink.Parsed;
 using Ink.Runtime;
+using Ink.UnityIntegration;
 using System.Collections;
 using System.Collections.Generic;
+using System.IO;
 using System.Text.RegularExpressions;
+using UnityEditor;
 using UnityEngine;
+
 
 public class Reader : MonoBehaviour
 {
-
     [SerializeField] private TextAsset _storyFile;
 
-    private Story _story;
+    private Ink.Runtime.Story _story;
 
     void Start()
     {
-        _story = new Story(_storyFile.text);
-        Read();
-    }
+        _story = new Ink.Runtime.Story(_storyFile.text);
+		
+		Read();
+
+		TestCondition();
+	}
+
+	void TestCondition()
+	{
+		string condition = "testFunction(\"hello\") && (T2 + 2 >= 7)";
+
+		InkParser p = new InkParser(condition);
+		Expression rootExpression = p.Expression();
+		Debug.Log($"{condition} => {_story.EvaluateAtRuntime(rootExpression)}");
+	}
 
     void Read()
     {
@@ -28,7 +45,7 @@ public class Reader : MonoBehaviour
 		{
 			for (int i = 0; i < _story.currentChoices.Count; ++i)
 			{
-				Choice choice = _story.currentChoices[i];
+				Ink.Runtime.Choice choice = _story.currentChoices[i];
 				Debug.Log($"{choice.condition} [{choice.isTrue}] {choice.text}");
 			}
 		}